Fix protocol parameters integer types (#1094)
* Fix protocol parameters integer types

* Fix block ID test

* Update sdk/src/types/block/protocol.rs


* Add doc

* Update sdk/src/types/block/protocol.rs

* Update sdk/src/types/block/protocol.rs


* Update sdk/src/types/block/protocol.rs


* Update sdk/src/types/block/protocol.rs


* Nits

* Apply suggestions from code review


* sad format noises

---------

+++ sdk/src/types/block/slot/index.rs
@@ -37,14 +37,14 @@ impl SlotIndex {
     /// Gets the [`EpochIndex`] of this slot.
     pub fn to_epoch_index(
         self,
-        slots_per_epoch_exponent_iter: impl Iterator<Item = (EpochIndex, u32)>,
+        slots_per_epoch_exponent_iter: impl Iterator<Item = (EpochIndex, u8)>,
     ) -> Result<EpochIndex, Error> {
         EpochIndex::from_slot_index(self, slots_per_epoch_exponent_iter)
     }
 
     /// Gets the slot index of a unix timestamp.
     /// Slots are counted starting from `1` with `0` being reserved for times before the genesis.
-    pub fn from_timestamp(timestamp: u64, genesis_unix_timestamp: u32, slot_duration_in_seconds: u8) -> SlotIndex {
+    pub fn from_timestamp(timestamp: u64, genesis_unix_timestamp: u64, slot_duration_in_seconds: u8) -> SlotIndex {
         timestamp
             .checked_sub(genesis_unix_timestamp as u64)
             .map(|diff| (diff / slot_duration_in_seconds as u64) + 1)
@@ -54,7 +54,7 @@ impl SlotIndex {
 
     /// Converts the slot index into the unix timestamp representing the beginning of the slot.
     /// Slot `0` will return the unix epoch.
-    pub fn to_timestamp(self, genesis_unix_timestamp: u32, slot_duration_in_seconds: u8) -> u64 {
+    pub fn to_timestamp(self, genesis_unix_timestamp: u64, slot_duration_in_seconds: u8) -> u64 {
         self.0
             .checked_sub(1)
             .map(|adjusted_slot| (adjusted_slot * slot_duration_in_seconds as u64) + genesis_unix_timestamp as u64)
